refactor(components): extract useHealthDashboardState hook
Further reduces HealthDashboardPage parent under the 500-line cap
per umbrella #290 requirement.

Part of #290
