+++ src/crosshook-native/src/components/pages/HealthDashboardPage.tsx
@@ -1,4 +1,4 @@
-import { Fragment, useDeferredValue, useEffect, useMemo, useRef, useState } from 'react';
+import { Fragment, useState } from 'react';
 import { useProfileContext } from '../../context/ProfileContext';
 import { useProfileHealthContext } from '../../context/ProfileHealthContext';
 import { useOfflineReadiness } from '../../hooks/useOfflineReadiness';
@@ -12,20 +12,14 @@ import type { AppRoute } from '../layout/Sidebar';
 import { MigrationReviewModal } from '../MigrationReviewModal';
 import { OfflineStatusBadge } from '../OfflineStatusBadge';
 import { CommunityImportHealthPanel } from './health-dashboard/CommunityImportHealthPanel';
-import type { CardTrend, SortDirection, SortField, StatusFilter } from './health-dashboard/constants';
-import { STATUS_RANK, VERSION_STATUS_RANK } from './health-dashboard/constants';
+import type { SortField } from './health-dashboard/constants';
 import { IssueBreakdownPanel, IssueDetailRow } from './health-dashboard/IssueBreakdownPanel';
 import { LauncherDriftPanel } from './health-dashboard/LauncherDriftPanel';
 import { RecentFailuresPanel } from './health-dashboard/RecentFailuresPanel';
 import { SkeletonCards, SummaryCard } from './health-dashboard/SummaryCards';
 import { SortArrow, TableToolbar } from './health-dashboard/TableControls';
-import {
-  categorizeIssue,
-  getVersionStatusColor,
-  getVersionStatusLabel,
-  mergeOfflineReadinessForRow,
-  offlineSortScore,
-} from './health-dashboard/utils';
+import { useHealthDashboardState } from './health-dashboard/useHealthDashboardState';
+import { getVersionStatusColor, getVersionStatusLabel, mergeOfflineReadinessForRow } from './health-dashboard/utils';
 
 export function HealthDashboardPage({ onNavigate }: { onNavigate?: (route: AppRoute) => void }) {
   const { summary, loading, error, batchValidate, cachedSnapshots, revalidateSingle, trendByName, staleInfoByName } =
@@ -44,16 +38,30 @@ export function HealthDashboardPage({ onNavigate }: { onNavigate?: (route: AppRo
   const offlineReadiness = useOfflineReadiness();
   const { checkVersionStatus } = useVersionCheck();
 
-  const [sortField, setSortField] = useState<SortField>('status');
-  const [sortDirection, setSortDirection] = useState<SortDirection>('desc');
-  const [statusFilter, setStatusFilter] = useState<StatusFilter>('all');
-  const [searchQuery, setSearchQuery] = useState('');
   const [expandedProfile, setExpandedProfile] = useState<string | null>(null);
   const [isMigrationModalOpen, setIsMigrationModalOpen] = useState(false);
   const [isVersionScanning, setIsVersionScanning] = useState(false);
   const [versionScanProgress, setVersionScanProgress] = useState<{ done: number; total: number } | null>(null);
 
-  const deferredSearch = useDeferredValue(searchQuery);
+  const {
+    sortField,
+    sortDirection,
+    statusFilter,
+    searchQuery,
+    setStatusFilter,
+    setSearchQuery,
+    allProfiles,
+    missingProtonCount,
+    filteredProfiles,
+    hasUnknownSentinel,
+    cachedSnapshotList,
+    recentFailures,
+    cardTrends,
+    ariaAnnouncement,
+    recheckPendingRef,
+    handleSortClick,
+    getAriaSortAttr,
+  } = useHealthDashboardState({ summary, loading, error, batchValidate, cachedSnapshots });
 
   async function handleFixNavigation(profileName: string) {
     await selectProfile(profileName);
@@ -87,194 +95,9 @@ export function HealthDashboardPage({ onNavigate }: { onNavigate?: (route: AppRo
     void batchValidate();
   }
 
-  function handleSortClick(field: SortField) {
-    if (field === sortField) {
-      setSortDirection((d) => (d === 'asc' ? 'desc' : 'asc'));
-    } else {
-      setSortField(field);
-      setSortDirection('asc');
-    }
-  }
-
   function handleRowClick(report: { name: string }) {
     setExpandedProfile((prev) => (prev === report.name ? null : report.name));
   }
-
-  const allProfiles = useMemo(() => {
-    return (summary?.profiles ?? []).filter((r) => r.name !== '<unknown>');
-  }, [summary?.profiles]);
-
-  const missingProtonCount = useMemo(() => {
-    return allProfiles.filter((r) => r.issues.some((issue) => categorizeIssue(issue) === 'missing_proton')).length;
-  }, [allProfiles]);
-
-  const filteredProfiles = useMemo(() => {
-    const term = deferredSearch.toLowerCase().trim();
-
-    let result = allProfiles;
-
-    if (statusFilter !== 'all') {
-      result = result.filter((r) => r.status === statusFilter);
-    }
-
-    if (term.length > 0) {
-      result = result.filter((r) => r.name.toLowerCase().includes(term));
-    }
-
-    result = result.slice().sort((a, b) => {
-      // Favorites always pin to top regardless of sort
-      const aFav = a.metadata?.is_favorite ?? false;
-      const bFav = b.metadata?.is_favorite ?? false;
-      if (aFav !== bFav) return aFav ? -1 : 1;
-
-      let cmp = 0;
-      switch (sortField) {
-        case 'name':
-          cmp = a.name.localeCompare(b.name);
-          break;
-        case 'status':
-          cmp = (STATUS_RANK[a.status] ?? 0) - (STATUS_RANK[b.status] ?? 0);
-          if (cmp === 0) cmp = a.name.localeCompare(b.name);
-          break;
-        case 'issues':
-          cmp = a.issues.length - b.issues.length;
-          break;
-        case 'last_success': {
-          const aSuccess = a.metadata?.last_success ?? '';
-          const bSuccess = b.metadata?.last_success ?? '';
-          cmp = aSuccess.localeCompare(bSuccess);
-          break;
-        }
-        case 'launch_method':
-          cmp = a.launch_method.localeCompare(b.launch_method);
-          break;
-        case 'failures':
-          cmp = (a.metadata?.failure_count_30d ?? 0) - (b.metadata?.failure_count_30d ?? 0);
-          break;
-        case 'favorite': {
-          const aFavSort = a.metadata?.is_favorite ? 1 : 0;
-          const bFavSort = b.metadata?.is_favorite ? 1 : 0;
-          cmp = aFavSort - bFavSort;
-          break;
-        }
-        case 'version_status': {
-          const aRank = VERSION_STATUS_RANK[a.metadata?.version_status ?? 'unknown'] ?? -1;
-          const bRank = VERSION_STATUS_RANK[b.metadata?.version_status ?? 'unknown'] ?? -1;
-          cmp = aRank - bRank;
-          break;
-        }
-        case 'offline_score': {
-          cmp =
-            offlineSortScore(a, offlineReadiness.reportForProfile(a.name)) -
-            offlineSortScore(b, offlineReadiness.reportForProfile(b.name));
-          if (cmp === 0) {
-            cmp = a.name.localeCompare(b.name);
-          }
-          break;
-        }
-        default:
-          cmp = 0;
-      }
-
-      return sortDirection === 'asc' ? cmp : -cmp;
-    });
-
-    return result;
-  }, [allProfiles, sortField, sortDirection, statusFilter, deferredSearch, offlineReadiness]);
-
-  const hasUnknownSentinel = (summary?.profiles ?? []).some((r) => r.name === '<unknown>');
-
-  const cachedSnapshotList = useMemo(() => {
-    return Object.values(cachedSnapshots)
-      .slice()
-      .sort((a, b) => {
-        const rankDiff = (STATUS_RANK[b.status] ?? 0) - (STATUS_RANK[a.status] ?? 0);
-        if (rankDiff !== 0) return rankDiff;
-        return a.profile_name.localeCompare(b.profile_name);
-      });
-  }, [cachedSnapshots]);
-
-  const recentFailures = useMemo(() => {
-    return (summary?.profiles ?? [])
-      .filter((r) => (r.metadata?.failure_count_30d ?? 0) > 0)
-      .slice()
-      .sort((a, b) => (b.metadata?.failure_count_30d ?? 0) - (a.metadata?.failure_count_30d ?? 0));
-  }, [summary?.profiles]);
-
-  const showLoadingCards = loading && !summary;
-  const isEmpty = !loading && summary?.total_count === 0;
-  const allHealthy =
-    summary !== null && summary.broken_count === 0 && summary.stale_count === 0 && summary.total_count > 0;
-
-  const cardTrends = useMemo<{ healthy: CardTrend; stale: CardTrend; broken: CardTrend }>(() => {
-    const snaps = Object.values(cachedSnapshots);
-    if (snaps.length === 0 || !summary) {
-      return { healthy: null, stale: null, broken: null };
-    }
-    const cachedHealthy = snaps.filter((s) => s.status === 'healthy').length;
-    const cachedStale = snaps.filter((s) => s.status === 'stale').length;
-    const cachedBroken = snaps.filter((s) => s.status === 'broken').length;
-    const healthyDiff = summary.healthy_count - cachedHealthy;
-    const staleDiff = summary.stale_count - cachedStale;
-    const brokenDiff = summary.broken_count - cachedBroken;
-    return {
-      healthy: healthyDiff > 0 ? 'up' : healthyDiff < 0 ? 'down' : null,
-      stale: staleDiff > 0 ? 'up' : staleDiff < 0 ? 'down' : null,
-      broken: brokenDiff > 0 ? 'up' : brokenDiff < 0 ? 'down' : null,
-    };
-  }, [cachedSnapshots, summary]);
-
-  const [ariaAnnouncement, setAriaAnnouncement] = useState('');
-  const recheckPendingRef = useRef(false);
-  const yButtonPrevRef = useRef(false);
-
-  useEffect(() => {
-    if (typeof window === 'undefined') return;
-
-    let rafId = 0;
-
-    const poll = () => {
-      const gamepad = navigator.getGamepads?.()[0];
-      if (gamepad) {
-        const yPressed = Boolean(gamepad.buttons[3]?.pressed);
-        const wasPressed = yButtonPrevRef.current;
-        if (yPressed && !wasPressed && !loading) {
-          recheckPendingRef.current = true;
-          void batchValidate();
-        }
-        yButtonPrevRef.current = yPressed;
-      }
-      rafId = window.requestAnimationFrame(poll);
-    };
-
-    rafId = window.requestAnimationFrame(poll);
-
-    return () => {
-      window.cancelAnimationFrame(rafId);
-    };
-  }, [loading, batchValidate]);
-
-  useEffect(() => {
-    if (loading) {
-      if (recheckPendingRef.current) {
-        setAriaAnnouncement('Checking all profiles...');
-      }
-      return;
-    }
-    if (!recheckPendingRef.current) {
-      return;
-    }
-    recheckPendingRef.current = false;
-    if (error) {
-      setAriaAnnouncement('');
-    } else if (summary) {
-      setAriaAnnouncement(
-        `Validation complete. ${summary.broken_count} broken, ${summary.stale_count} stale, ${summary.healthy_count} healthy.`
-      );
-    } else {
-      setAriaAnnouncement('Validation complete.');
-    }
-  }, [loading, error, summary]);
 
   function handleRetry() {
     if (error) {
@@ -286,11 +109,6 @@ export function HealthDashboardPage({ onNavigate }: { onNavigate?: (route: AppRo
   function handleRecheck() {
     recheckPendingRef.current = true;
     void batchValidate();
-  }
-
-  function getAriaSortAttr(field: SortField): 'ascending' | 'descending' | 'none' {
-    if (sortField !== field) return 'none';
-    return sortDirection === 'asc' ? 'ascending' : 'descending';
   }
 
   function renderSortHeader(field: SortField, label: string) {
@@ -308,6 +126,11 @@ export function HealthDashboardPage({ onNavigate }: { onNavigate?: (route: AppRo
       </th>
     );
   }
+
+  const showLoadingCards = loading && !summary;
+  const isEmpty = !loading && summary?.total_count === 0;
+  const allHealthy =
+    summary !== null && summary.broken_count === 0 && summary.stale_count === 0 && summary.total_count > 0;
 
   return (
     <div className="crosshook-page-scroll-shell crosshook-page-scroll-shell--health">
